refactor: extract CoachingData to separate file
