+++ socket-service/src/main/kotlin/com/parkit/socket/dto/CoachingData.kt
@@ -1,0 +1,9 @@
+package com.parkit.socket.dto
+
+data class CoachingData(
+    val sensorId: String,
+    val distance: Double,
+    val warningLevel: String,
+    val message: String,
+    val timestamp: Long
+)
